1 unit test fails

+++ sorting_algorithms.h
@@ -20,7 +20,7 @@
 *** Basically two functions to sort arrays of integer elements. Furthermore, a function
 *** to initialize an array with a number of random elements.
 */
-	void init_random(int small_array[], int SMALL);
-	void bubble_sort(int small_array[], int SMALL);
-  void insertion_sort(int small_array[], int SMALL);
+	void init_random(int small_array[], unsigned long SMALL);
+	void bubble_sort(int small_array[], unsigned long SMALL);
+  void insertion_sort(int small_array[], unsigned long SMALL);
 #endif
